@@ -24,21 +24,27 @@
 #  index_districts_on_nces_id  (nces_id) UNIQUE
 #
 class District < ApplicationRecord
+  include Subscriber
 
   validates :name, presence: true
 
   has_many :schools
-  has_many :district_admins, class_name: 'DistrictAdmin', dependent: :destroy
+  has_many :district_admins, dependent: :destroy
   has_many :admins, through: :district_admins, source: :user
+  has_many :district_subscriptions
+  has_many :subscriptions, through: :district_subscriptions
 
-  scope :by_name, ->(name) { where('lower(name) LIKE ?', "%#{name.downcase}%") }
-  scope :by_city, ->(city) { where('lower(city) LIKE ?', "%#{city.downcase}%") }
-  scope :by_state, ->(state) { where(:state => state.upcase) }
-  scope :by_zipcode, ->(zipcode) { where(:zipcode => zipcode) }
-  scope :by_nces_id, ->(nces_id) { where(:nces_id => nces_id) }
+  scope :by_name, ->(name) { where('name ILIKE ?', "%#{name}%") }
+  scope :by_city, ->(city) { where('city ILIKE ?', "%#{city}%") }
+  scope :by_state, ->(state) { where(state: state.upcase) }
+  scope :by_zipcode, ->(zipcode) { where(zipcode: zipcode) }
+  scope :by_nces_id, ->(nces_id) { where(nces_id: nces_id) }
+
+  def attach_subscription(subscription)
+    district_subscriptions.create(subscription: subscription)
+  end
 
   def total_invoice
     schools.sum { |s| s&.subscription&.payment_amount || 0 } / 100.0
   end
-
 end
