@@ -2,7 +2,6 @@
 
 class Cms::SchoolsController < Cms::CmsController
   before_action :signed_in!
-
   before_action :text_search_inputs, only: [:index, :search]
   before_action :set_school, only: [:new_subscription, :edit_subscription, :show, :complete_sales_stage]
   before_action :subscription_data, only: [:new_subscription, :edit_subscription]
@@ -73,7 +72,11 @@ class Cms::SchoolsController < Cms::CmsController
   end
 
   def new_subscription
-    @subscription = Subscription.new(start_date: Subscription.redemption_start_date(@school), expiration: Subscription.default_expiration_date(@school))
+    @subscription = Subscription.new(
+      account_type: Subscription::SCHOOL_PAID,
+      start_date: Subscription.redemption_start_date(@school),
+      expiration: Subscription.default_expiration_date(@school)
+    )
   end
 
   def edit_subscription
@@ -353,5 +356,14 @@ class Cms::SchoolsController < Cms::CmsController
 
   def fallback_location
     cms_school_path(params[:id].to_i)
+  end
+
+  private def subscription_data
+    @premium_types = Subscription::OFFICIAL_SCHOOL_TYPES
+    @subscription_payment_methods = Subscription::CMS_PAYMENT_METHODS
+
+    return if @school.nil? || @school.alternative?
+
+    @school_users = @school.users.select(:id, :email, :name)
   end
 end
