@@ -15,7 +15,6 @@ class PagesController < ApplicationController
   NUMBER_OF_SCHOOLS = "NUMBER_OF_SCHOOLS"
   NUMBER_OF_LOW_INCOME_SCHOOLS = "NUMBER_OF_LOW_INCOME_SCHOOLS"
   OPEN_POSITIONS = Configs[:careers][:open_positions]
-  SHOW_SCHOOL_BUY_NOW_BUTTON_APP_SETTING = 'show_school_buy_now_button'
 
   def home
     if signed_in?
@@ -405,14 +404,13 @@ class PagesController < ApplicationController
   def premium
     @user_is_eligible_for_new_subscription= current_user&.eligible_for_new_subscription?
     @user_is_eligible_for_trial = current_user&.subscriptions&.none?
-
-    @user_has_school = !!current_user&.school && ['home school', 'us higher ed', 'international', 'other', 'not listed'].exclude?(current_user&.school&.name)
-    @user_belongs_to_school_that_has_paid = current_user&.school ? Subscription.school_or_user_has_ever_paid?(current_user&.school) : false
+    @user_has_school = !!current_user&.school && !current_user.school.alternative?
+    @user_belongs_to_school_that_has_paid = !!current_user&.school&.ever_paid_for_subscription?
     @customer_email = current_user&.email
-    @school_ids = [current_user&.school&.id].to_json
+    @associated_schools = current_user&.associated_schools || []
+    @eligible_schools = @associated_schools.filter { |s| s.subscription.nil? }
     @stripe_school_plan = PlanSerializer.new(Plan.stripe_school_plan).as_json
     @stripe_teacher_plan = PlanSerializer.new(Plan.stripe_teacher_plan).as_json
-    @show_school_buy_now = AppSetting.enabled?(name: SHOW_SCHOOL_BUY_NOW_BUTTON_APP_SETTING, user: current_user)
 
     @diagnostic_activity_count =
       Activity.where(
