@@ -101,12 +101,11 @@ class SchoolsController < ApplicationController
             name: school_params[:school_id_or_type]
           )
         end
-        school_user = SchoolsUsers.find_or_create_by(
-          user_id: current_user.id,
-          school_id: school.id
-        )
+        school_user = SchoolsUsers.find_by(user_id: current_user.id) || SchoolsUsers.new(user_id: current_user.id)
+        school_user.school_id = school.id
+        school_user.save
         find_or_create_checkbox('Add School', current_user)
-        render json: {}
+        render json: { school: school, subscription: school.subscription }
       }
     end
   end
