@@ -561,9 +561,9 @@ class ActivitySession < ApplicationRecord
 
     case sort[:field]
     when 'activity_classification_name'
-      "activity_classifications.name #{order}, #{last_name} #{order}"
+      Arel.sql("activity_classifications.name #{order}, #{last_name} #{order}")
     when 'student_name'
-      "#{last_name} #{order}, users.name #{order}"
+      Arel.sql("#{last_name} #{order}, users.name #{order}")
     when 'completed_at'
       "activity_sessions.completed_at #{order}"
     when 'activity_name'
