@@ -20,34 +20,29 @@ class SchoolSubscription < ApplicationRecord
   belongs_to :school
   belongs_to :subscription
   after_commit :update_schools_users
-  after_commit :attach_district_admins
   after_create :send_premium_emails
 
   def update_schools_users
     return unless school&.users
 
-    school.users.each do |u|
-      UserSubscription.create_user_sub_from_school_sub_if_they_do_not_have_that_school_sub(u.id, subscription_id)
+    school.users.each do |user|
+      UserSubscription.create_user_sub_from_school_sub_if_they_do_not_have_that_school_sub(user, subscription)
     end
   end
 
+  # rubocop:disable Metrics/CyclomaticComplexity
   def send_premium_emails
     return unless school&.users
 
     if Rails.env.production?
-      school.users.each do |u|
-        PremiumSchoolSubscriptionEmailWorker.perform_async(u.id)
+      school.users.each do |user|
+        PremiumSchoolSubscriptionEmailWorker.perform_async(user.id)
       end
     else
-      school.users.each do |u|
-        PremiumSchoolSubscriptionEmailWorker.perform_async(u.id) if u.email.match('quill.org')
+      school.users.each do |user|
+        PremiumSchoolSubscriptionEmailWorker.perform_async(user.id) if user.email&.match('quill.org')
       end
     end
   end
-
-  def attach_district_admins
-    school&.district&.district_admins&.each do |da|
-      da.attach_to_subscribed_schools
-    end
-  end
+  # rubocop:enable Metrics/CyclomaticComplexity
 end
