@@ -34,11 +34,13 @@ require 'new_relic/agent'
 class Subscription < ApplicationRecord
   class RenewalNilStripeCustomer < StandardError; end
 
+  has_many :credit_transactions, as: :source
+  has_many :district_subscriptions
+  has_many :districts, through: :district_subscriptions
+  has_many :school_subscriptions
+  has_many :schools, through: :school_subscriptions
   has_many :user_subscriptions
   has_many :users, through: :user_subscriptions
-  has_many :school_subscriptions
-  has_many :credit_transactions, as: :source
-  has_many :schools, through: :school_subscriptions
   belongs_to :purchaser, class_name: "User"
   belongs_to :subscription_type
   belongs_to :plan, optional: true
@@ -65,19 +67,22 @@ class Subscription < ApplicationRecord
     TEACHER_PAID,
     PREMIUM_CREDIT,
     CB_LIFETIME_SUBSCRIPTION_TYPE
-  ]
+  ].freeze
 
   OFFICIAL_FREE_TYPES = [
     SCHOOL_SPONSORED_FREE,
     TEACHER_SPONSORED_FREE,
     TEACHER_TRIAL
-  ]
+  ].freeze
 
   OFFICIAL_SCHOOL_TYPES = [
-    SCHOOL_DISTRICT_PAID,
     SCHOOL_PAID,
     SCHOOL_SPONSORED_FREE
-  ]
+  ].freeze
+
+  OFFICIAL_DISTRICT_TYPES = [
+    SCHOOL_DISTRICT_PAID
+  ].freeze
 
   OFFICIAL_TEACHER_TYPES = [
     TEACHER_PAID,
@@ -85,12 +90,10 @@ class Subscription < ApplicationRecord
     TEACHER_SPONSORED_FREE,
     TEACHER_TRIAL,
     CB_LIFETIME_SUBSCRIPTION_TYPE
-  ]
+  ].freeze
 
   ALL_OFFICIAL_TYPES = OFFICIAL_PAID_TYPES + OFFICIAL_FREE_TYPES
   TRIAL_TYPES = [TEACHER_TRIAL]
-
-  SCHOOL_RENEWAL_PRICE = 90000
 
   TYPES_HASH = {
     trial: TRIAL_TYPES,
@@ -98,17 +101,18 @@ class Subscription < ApplicationRecord
     school: OFFICIAL_SCHOOL_TYPES
   }
 
-  # 2/27 - for now every school is 90000 cents, whether they are renewing or re-signing up.
-  SCHOOL_FIRST_PURCHASE_PRICE = SCHOOL_RENEWAL_PRICE
-  TEACHER_PRICE = 8000
-  ALL_PRICES = [TEACHER_PRICE, SCHOOL_RENEWAL_PRICE]
   PAYMENT_METHODS = [
     INVOICE_PAYMENT_METHOD = 'Invoice',
     CREDIT_CARD_PAYMENT_METHOD = 'Credit Card',
     PREMIUM_CREDIT_PAYMENT_METHOD = 'Premium Credit'
-  ]
+  ].freeze
 
-  ALL_TYPES = OFFICIAL_FREE_TYPES.dup.concat(OFFICIAL_PAID_TYPES)
+  CMS_PAYMENT_METHODS = [
+    INVOICE_PAYMENT_METHOD,
+    PREMIUM_CREDIT_PAYMENT_METHOD
+  ].freeze
+
+  ALL_TYPES = OFFICIAL_FREE_TYPES.dup.concat(OFFICIAL_PAID_TYPES).freeze
 
   validates :stripe_invoice_id, allow_blank: true, stripe_uid: { prefix: :in }
 
@@ -124,6 +128,38 @@ class Subscription < ApplicationRecord
   scope :not_stripe, -> { where(stripe_invoice_id: nil) }
   scope :started, -> { where("start_date <= ?", Date.current) }
   scope :paid_with_card, -> { where.not(stripe_invoice_id: nil).or(where(payment_method: 'Credit Card')) }
+
+  def self.create_and_attach_subscriber(subscription_attrs, subscriber)
+    if !subscription_attrs[:expiration]
+      case subscription_attrs[:account_type]
+      when TEACHER_TRIAL
+        subscription_attrs = subscription_attrs.merge(set_trial_expiration_and_start_date(subscriber))
+      when CB_LIFETIME_SUBSCRIPTION_TYPE
+        subscription_attrs = subscription_attrs.merge(set_cb_lifetime_expiration_and_start_date)
+      else
+        subscription_attrs = subscription_attrs.merge(set_premium_expiration_and_start_date(subscriber))
+      end
+    end
+
+    ActiveRecord::Base.transaction(requires_new: true) do
+      subscriber.subscriptions.each do |existing_subscription|
+        existing_subscription.update!(recurring: false)
+        existing_subscription.stripe_cancel_at_period_end
+      end
+
+      subscription = Subscription.create!(subscription_attrs)
+      subscriber.attach_subscription(subscription)
+      subscription
+    end
+  end
+
+  def premium_types
+    return OFFICIAL_DISTRICT_TYPES if district_subscriptions.exists?
+    return OFFICIAL_SCHOOL_TYPES if school_subscriptions.exists?
+    return OFFICIAL_TEACHER_TYPES if user_subscriptions.exists?
+
+    []
+  end
 
   def is_trial?
     account_type && TRIAL_TYPES.include?(account_type)
@@ -143,20 +179,8 @@ class Subscription < ApplicationRecord
     update(purchaser_id: new_purchaser_id)
   end
 
-  def renewal_price
-    if schools.any?
-      SCHOOL_RENEWAL_PRICE
-    else
-      TEACHER_PRICE
-    end
-  end
-
-  def self.create_with_user_join user_id, attributes
-    create_with_school_or_user_join user_id, 'User', attributes
-  end
-
   def school_subscription?
-    SchoolSubscription.where(subscription_id: id).limit(1).exists?
+    SchoolSubscription.exists?(subscription_id: id)
   end
 
   def self.account_types
@@ -164,12 +188,10 @@ class Subscription < ApplicationRecord
   end
 
   def credit_user_and_de_activate
-    return if school_subscriptions.ids.any? || user_subscriptions.ids.count > 1
+    return if district_subscriptions.ids.any? || school_subscriptions.ids.any? || user_subscriptions.ids.count > 1
 
     update(de_activated_date: Date.current, recurring: false)
     stripe_cancel_at_period_end
-    # subtract later of start date or today's date from expiration date to calculate amount to credit
-    # amount_to_credit = self.expiration - [self.start_date, Date.current].max
     amount_to_credit = expiration - start_date
     CreditTransaction.create(user_id: user_subscriptions.first.user_id, amount: amount_to_credit.to_i, source: self)
   end
@@ -182,68 +204,18 @@ class Subscription < ApplicationRecord
       .recurring
   end
 
-  def self.expired_today_or_previously_and_not_recurring
-    Subscription
-      .expired
-      .not_de_activated
-      .not_recurring
+  def self.redemption_start_date(subscriber)
+    subscriber&.subscriptions&.active&.first&.expiration || Date.current
   end
 
-  def self.school_or_user_has_ever_paid?(school_or_user)
-    (OFFICIAL_PAID_TYPES & school_or_user.subscriptions.pluck(:account_type)).present?
-  end
-
-  def self.new_school_premium_sub(school, user)
-    today = Date.current
-    expiration = school_or_user_has_ever_paid?(school) ? (today + 1.year) : promotional_dates[:expiration]
-    new(expiration: expiration, start_date: today, account_type: 'School Paid', recurring: true, purchaser_id: user.id)
-  end
-
-  def self.give_school_premium_if_charge_succeeds(school, user)
-    school_premium_sub = new_school_premium_sub(school, user)
-    school_premium_sub.save_if_charge_succeeds('school', school)
-    if school_premium_sub.new_record?
-      false
-    else
-      SchoolSubscription.create(school: school, subscription: school_premium_sub)
-      school_premium_sub
-    end
-  end
-
-  def self.redemption_start_date(school_or_user)
-    school_or_user&.subscriptions&.active&.first&.expiration || Date.current
-  end
-
-  def self.default_expiration_date(school_or_user)
-    last_subscription = school_or_user.subscriptions.active.first
+  def self.default_expiration_date(subscriber)
+    last_subscription = subscriber.subscriptions.active.first
     if last_subscription.present?
-      redemption_start_date(school_or_user) + 1.year
-    elsif school_or_user.instance_of?(School)
+      redemption_start_date(subscriber) + 1.year
+    elsif subscriber.promotional_dates?
       promotional_dates[:expiration]
     else
       Date.current + 1.year
-    end
-  end
-
-  def save_if_charge_succeeds(premium_type, school=nil)
-    raise "an incorrect premium type #{premium_type} was passed" unless premium_type == 'school'
-
-    charge = charge_user_for_school_premium(school)
-    payment_amount = SCHOOL_RENEWAL_PRICE
-
-    if charge[:status] == 'succeeded'
-      self.payment_method = 'Credit Card'
-      self.payment_amount = payment_amount
-      save!
-      self
-    else
-      nil
-    end
-  end
-
-  def detach_district_admins
-    schools.each do |school|
-      school.detach_from_existing_district_admins(school.district)
     end
   end
 
@@ -271,41 +243,24 @@ class Subscription < ApplicationRecord
     end
   end
 
-  def self.update_todays_expired_school_subscriptions
-    expired_today_or_previously_and_not_recurring.where(account_type: OFFICIAL_SCHOOL_TYPES).each do |subscription|
-      subscription.detach_district_admins
-    end
-  end
-
   def self.promotional_dates
     today = Date.current
-    # available to users who have never paid before
-    # if today's month is before july, it expires end of June, else December
     exp_month_and_day = today.month < 7 ? "30-6" : "31-12"
 
     { start_date: today, expiration: Date::strptime("#{exp_month_and_day}-#{today.year + 1}","%d-%m-%Y") }
   end
 
-  protected def charge_user_for_school_premium(school)
-    return unless purchaser&.stripe_customer?
-
-    Stripe::Charge.create(amount: SCHOOL_FIRST_PURCHASE_PRICE, currency: 'usd', customer: purchaser.stripe_customer_id)
-  end
-
-  def self.set_premium_expiration_and_start_date(school_or_user)
+  def self.set_premium_expiration_and_start_date(subscriber)
     today = Date.current
 
-    if !Subscription.school_or_user_has_ever_paid?(school_or_user) && school_or_user.instance_of?(School)
+    if !subscriber.ever_paid_for_subscription? && subscriber.promotional_dates?
       # We end their trial if they have one
-      school_or_user.subscription&.update(de_activated_date: today)
-      # Then they get the promotional subscription
+      subscriber.subscription&.update(de_activated_date: today)
       promotional_dates
-    elsif school_or_user.subscription
-      # Expire one year later, start at end of sub
-      old_sub = school_or_user.subscription
+    elsif subscriber.subscription
+      old_sub = subscriber.subscription
       { expiration: old_sub.expiration + 1.year, start_date: old_sub.expiration }
     else
-      # sub lasts one year from Date.current
       { expiration: today + 1.year, start_date: today }
     end
   end
@@ -340,38 +295,6 @@ class Subscription < ApplicationRecord
     self.start_date = Date.current
   end
 
-  def self.create_with_school_or_user_join(school_or_user_id, type, attributes)
-    type = type.capitalize
-    # since we're constantizing the type, need to make sure it is capitalized if not already
-    school_or_user = type.constantize.find school_or_user_id
-    # get school or user object
-    if !attributes[:expiration]
-      # if there is no expiration, either give them a trial one or a premium one
-      # TODO: 'subscription type spot'
-      if attributes[:account_type]&.downcase == 'teacher trial'
-        attributes = attributes.merge(Subscription.set_trial_expiration_and_start_date(school_or_user))
-      elsif attributes[:account_type] == CB_LIFETIME_SUBSCRIPTION_TYPE
-        attributes = attributes.merge(Subscription.set_cb_lifetime_expiration_and_start_date)
-      else
-        attributes = attributes.merge(Subscription.set_premium_expiration_and_start_date(school_or_user))
-      end
-    end
-
-    school_or_user.subscriptions.each do |subscription|
-      subscription.update!(recurring: false)
-      subscription.stripe_cancel_at_period_end
-    end
-
-    subscription = Subscription.create!(attributes)
-
-    h = {}
-    h["#{type.downcase}_id".to_sym] = school_or_user_id
-    h[:subscription_id] = subscription.id
-    "#{type}Subscription".constantize.create(h)
-
-    subscription
-  end
-
   def last_four
     stripe_subscription&.last_four || purchaser&.last_four
   end
@@ -397,7 +320,7 @@ class Subscription < ApplicationRecord
 
   def renewal_stripe_price_id
     return STRIPE_TEACHER_PLAN_PRICE_ID if [TEACHER_PAID, TEACHER_TRIAL].include?(account_type)
-    # can get cleaned up when we unify account types vs plan names, this covers the existing bases
+    # TODO: can get cleaned up when we unify account types vs plan names, this covers the existing bases
     return STRIPE_SCHOOL_PLAN_PRICE_ID if account_type == SCHOOL_PAID
     return STRIPE_SCHOOL_PLAN_PRICE_ID if account_type == Plan::STRIPE_SCHOOL_PLAN
   end
