@@ -5,6 +5,9 @@ import BasicPricingMini from './premium_minis/basic_pricing_mini.jsx';
 import TeacherPricingMini from './premium_minis/teacher_pricing_mini.jsx';
 import SchoolPricingMini from './premium_minis/school_pricing_mini.jsx';
 
+const MOBILE_WIDTH = 991
+const VERTICAL_INTERSECTION_OF_PREMIUM_PRICING_ROW_AND_TABLE = 337
+
 export default class PremiumPricingMinisRow extends React.Component {
   state = {
     subscriptionType: null,
@@ -19,9 +22,9 @@ export default class PremiumPricingMinisRow extends React.Component {
 
   listenScrollEvent = e => {
     const { isScrolled, } = this.state
-    if (window.scrollY > 337 && !isScrolled) {
+    if (window.scrollY > VERTICAL_INTERSECTION_OF_PREMIUM_PRICING_ROW_AND_TABLE && !isScrolled && window.innerWidth > MOBILE_WIDTH) {
       this.setState({isScrolled: true})
-    } else if (window.scrollY < 337 && isScrolled) {
+    } else if (window.scrollY < VERTICAL_INTERSECTION_OF_PREMIUM_PRICING_ROW_AND_TABLE && isScrolled) {
       this.setState({isScrolled: false})
     }
   }
@@ -31,11 +34,10 @@ export default class PremiumPricingMinisRow extends React.Component {
       diagnosticActivityCount,
       lessonsActivityCount,
       independentPracticeActivityCount,
-      schoolBuyNowButton,
-      showSchoolBuyNow,
       stripeSchoolPlan,
       stripeTeacherPlan,
       teacherBuyNowButton,
+      onClickPurchasingOptions,
     } = this.props
 
     const { userIsSignedIn, isScrolled, } = this.state
@@ -64,11 +66,10 @@ export default class PremiumPricingMinisRow extends React.Component {
               premiumFeatureData={premiumFeatureData}
             />
             <SchoolPricingMini
-              buyNowButton={schoolBuyNowButton}
+              handleClickPurchasingOptions={onClickPurchasingOptions}
               plan={stripeSchoolPlan.plan}
               premiumFeatureData={premiumFeatureData}
               showBadges={!isScrolled}
-              showSchoolBuyNow={showSchoolBuyNow}
             />
           </div>
         </div>
