@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { requestPost } from '../../../../modules/request';
 
 export const StripeSubscriptionCheckoutSessionButton = ({
@@ -18,9 +19,7 @@ export const StripeSubscriptionCheckoutSessionButton = ({
       alert('You must be logged in to activate Premium.')
     } else if (!userIsEligibleForNewSubscription) {
       alert(
-        `You have an active subscription and cannot buy premium now. If your subscription is a school subscription,
-        you may buy Premium when it expires. If your subscription is a teacher one, please turn on recurring payments
-        and we will renew it automatically when your subscription ends.`
+        "You have an active subscription and cannot buy premium now. If your subscription is a school subscription, you may buy Premium when it expires. If your subscription is a teacher one, please turn on recurring payments and we will renew it automatically when your subscription ends."
       )
     } else {
       const path = '/stripe_integration/subscription_checkout_sessions'
