@@ -2,6 +2,7 @@ import React from 'react';
 import { SingleDatePicker } from 'react-dates'
 import _ from 'lodash';
 import request from 'request';
+import { DataTable } from '../../Shared/index'
 import ItemDropdown from '../components/general_components/dropdown_selectors/item_dropdown.jsx';
 import getAuthToken from '../components/modules/get_auth_token';
 import moment from 'moment';
@@ -10,7 +11,7 @@ export default class EditOrCreateSubscription extends React.Component {
   constructor(props) {
     super(props);
 
-    const { subscription, current_subscription, } = props
+    const { schools, subscription, current_subscription, } = props
 
     const defaultSubscription = subscription || current_subscription
 
@@ -21,31 +22,9 @@ export default class EditOrCreateSubscription extends React.Component {
     this.state = {
       subscription: defaultSubscription,
       firstFocused: false,
-      secondFocused: false
-    };
-  }
-
-  getMatchingUserFromSchoolsUsersByEmail = (newSub, e) => {
-    const { schoolsUsers, } = this.props
-    const matchingSchoolUser = schoolsUsers && schoolsUsers.find(u => u.email === e.target.value);
-    if (matchingSchoolUser) {
-      newSub.purchaser_id = matchingSchoolUser.id;
+      secondFocused: false,
+      schools
     }
-  }
-
-  getMatchingUserFromSchoolsUsersById = (newSub, e) => {
-    const { schoolsUsers, } = this.props
-    const matchingSchoolUser = schoolsUsers && schoolsUsers.find(u => u.id === e);
-    if (matchingSchoolUser) {
-      newSub.purchaser_email = matchingSchoolUser.email;
-    }
-  }
-
-  changeAccountType = (e) => {
-    const { subscription, } = this.state
-    const newSub = Object.assign({}, subscription);
-    newSub.account_type = e;
-    this.setState({ subscription: newSub, });
   }
 
   changePaymentMethod = (e) => {
@@ -70,6 +49,70 @@ export default class EditOrCreateSubscription extends React.Component {
     if (user) {
       return <span className="green-text text-green" onClick={this.handleChangePurchaserInfoToTeacherInfo}>Same As Teacher Info</span>;
     }
+  }
+
+  createdAt = () => {
+    const { subscription } = this.state
+    const { created_at } = subscription
+
+    if (!created_at) { return }
+
+    return (
+      <React.Fragment>
+        <label>Created At:</label>
+        <span>{created_at}</span>
+        <br />
+      </React.Fragment>
+    )
+  }
+
+  endDate = () => {
+    const { promoExpiration } = this.props
+    const { secondFocused, subscription } = this.state
+
+    return (
+      <React.Fragment>
+        <label htmlFor="">End Date</label>
+        <p>
+          If this a school's first paid subscription, the default end date is {promoExpiration}.
+          This value just stated will update automatically depending on the time of year.
+        </p>
+        <SingleDatePicker
+          date={subscription.expiration ? moment(subscription.expiration) : null}
+          focused={secondFocused}
+          id="date-picker"
+          inputIconPosition="after"
+          navNext="›"
+          navPrev="‹"
+          numberOfMonths={1}
+          onDateChange={this.handleExpirationDateChange}
+          onFocusChange={() => this.setState({ secondFocused: !secondFocused })}
+        />
+      </React.Fragment>
+    )
+  }
+
+  getMatchingUserFromSchoolsUsersByEmail = (newSub, e) => {
+    const { schoolsUsers, } = this.props
+    const matchingSchoolUser = schoolsUsers && schoolsUsers.find(u => u.email === e.target.value);
+    if (matchingSchoolUser) {
+      newSub.purchaser_id = matchingSchoolUser.id;
+    }
+  }
+
+  getMatchingUserFromSchoolsUsersById = (newSub, e) => {
+    const { schoolsUsers, } = this.props
+    const matchingSchoolUser = schoolsUsers && schoolsUsers.find(u => u.id === e);
+    if (matchingSchoolUser) {
+      newSub.purchaser_email = matchingSchoolUser.email;
+    }
+  }
+
+  handleChangeAccountType = (e) => {
+    const { subscription, } = this.state
+    const newSub = Object.assign({}, subscription);
+    newSub.account_type = e;
+    this.setState({ subscription: newSub, });
   }
 
   handleChangePurchaserInfoToTeacherInfo = () => {
@@ -117,6 +160,51 @@ export default class EditOrCreateSubscription extends React.Component {
     this.setState({ subscription: newSub, });
   }
 
+  handleSubmit = () => {
+    const { view, } = this.props
+    const submitVars = this.submitVars()
+    const that = this
+
+    request[submitVars.httpVerb](
+      {
+        url: submitVars.urlString,
+        json: submitVars.data,
+      },
+      (e, httpResponse, _body) => {
+        if (httpResponse.statusCode === 200) {
+          alert('Subscription was saved');
+          if (view === 'new') {
+            // switch to the edit view after submission
+            window.location = window.location.href.replace('new', 'edit')
+          }
+        } else {
+          alert('There was an error. Please try again and contact a dev if you continue to get this warning.')
+        }
+      }
+    )
+  }
+
+  paymentInformation = () => {
+    const { subscriptionPaymentMethods  } = this.props
+    const { subscription } = this.state
+    const subscriptionPaymentOptions = subscriptionPaymentMethods.concat('N/A')
+
+    return (
+      <React.Fragment>
+        <h2>Payment Information</h2>
+        <label>Payment Method</label>
+        <ItemDropdown
+          callback={this.changePaymentMethod}
+          className="subscription-dropdown"
+          items={subscriptionPaymentOptions}
+          selectedItem={subscription.payment_method || 'N/A'}
+        />
+        <label>Purchase Amount (dollar value as integer -- no decimal or symbol)</label>
+        <input onChange={this.handlePaymentAmountChange} type="text" value={subscription.payment_amount / 100} />
+      </React.Fragment>
+    )
+  }
+
   purchaserFromSchool = () => {
     const { subscription, } = this.state
     const { schoolsUsers, } = this.props
@@ -135,13 +223,58 @@ export default class EditOrCreateSubscription extends React.Component {
     }
   }
 
-  recurringIfCreditCard = () => {
+  period = () => {
+    const { view } = this.props
+
+    return (
+      <React.Fragment>
+        <h2>Period {view === 'edit' ? <span className="warning text-red">-- These should not be edited without good reason!</span> : null}</h2>
+        {this.startDate()}
+        {this.endDate()}
+      </React.Fragment>
+    )
+  }
+
+  premiumStatus = () => {
+    const { premiumTypes } = this.props
+    const { subscription } = this.state
+
+    return (
+      <React.Fragment>
+        <label>Premium Status</label>
+        <ItemDropdown
+          callback={this.changeAccountType}
+          className="subscription-dropdown"
+          items={premiumTypes}
+          selectedItem={subscription.account_type || ''}
+        />
+      </React.Fragment>
+    )
+  }
+
+  purchaserInformation = () => {
+    const { subscription } = this.state
+
+    return (
+      <React.Fragment>
+        <h2>Purchaser Information</h2>
+        {this.purchaserFromSchool()}
+        <label>Purchaser Email</label>
+        <p>If the purchaser is not in the school and you see a school dropdown, select &#39;None&#39; and put in the purchasers email.</p>
+        <input onChange={this.handlePurchaserEmailChange} type="text" value={subscription.purchaser_email} />
+        <br />
+        {this.changeToPurchaserInfo()}
+      </React.Fragment>
+    )
+  }
+
+  recurring = () => {
     const { subscription, } = this.state
 
     if (subscription.payment_method === 'Credit Card') {
       return (
         <label>
-                Recurring:
+          Recurring:
           <input
             checked={subscription.recurring}
             onChange={this.handleRecurringChange}
@@ -152,98 +285,40 @@ export default class EditOrCreateSubscription extends React.Component {
     }
   }
 
-  submit = () => {
-    const { view, } = this.props
-    const submitVars = this.submitVars();
-    const that = this;
-    request[submitVars.httpVerb]({
-      url: submitVars.urlString,
-      json: submitVars.data,
-    },
-    (e, httpResponse, body) => {
-      if (httpResponse.statusCode === 200) {
-        alert('Subscription was saved');
-        if (view === 'new') {
-          // switch to the edit view after submission
-          window.location = window.location.href.replace('new', 'edit');
-        }
-      } else {
-        alert('There was an error. Please try again and contact a dev if you continue to get this warning.');
-      }
-    });
-  }
+  schoolSubscriptions = () => {
+    const { district, schools } = this.props
+    const headers = [{width: '510px', name: 'School', attribute: 'name'}]
 
-  submitConfirmation = () => {
-    const { school, } = this.props
-    if (school) {
-      if (confirm("You know you are about to add/edit an entire school's subscription, right?")) {
-        this.submit();
-      } else {
-        alert('submission canceled');
-      }
-    }
-  }
+    if (!district || !schools || !schools.length) { return }
 
-  submitVars = () => {
-    const { subscription, } = this.state
-    const { view, schoolOrUser, user, school } = this.props
-
-    const varsObj = { data: { subscription: subscription, authenticity_token: getAuthToken(), }, urlString: `${process.env.DEFAULT_URL}/cms/subscriptions`, };
-    if (view === 'edit') {
-      varsObj.httpVerb = 'put';
-      varsObj.urlString += `/${subscription.id}`;
-    } else {
-      // we are creating
-      varsObj.data.school_or_user = schoolOrUser;
-      varsObj.data.school_or_user_id = schoolOrUser === 'school' ? school.id : user.id
-      varsObj.httpVerb = 'post';
-    }
-    return varsObj;
-  }
-
-  render() {
-    const { user, school, view, premiumTypes, subscriptionPaymentMethods, promoExpiration, } = this.props
-    const { subscription, firstFocused, secondFocused } = this.state
-
-    const schoolOrUser = school || user || null;
-    const submitAction = school ? this.submitConfirmation : this.submit;
-    const subscriptionPaymentOptions = subscriptionPaymentMethods.concat('N/A')
     return (
-      <div className="cms-subscription">
-        <h1>{view === 'edit' ? 'Edit' : 'New'} Subscription: {_.get(schoolOrUser, 'name')}</h1>
-        <h2>Subscription Information</h2>
-        <label>Premium Status</label>
-        <ItemDropdown
-          callback={this.changeAccountType}
-          className="subscription-dropdown"
-          items={premiumTypes}
-          selectedItem={subscription.account_type || ''}
+      <React.Fragment>
+        <h2>School Subscriptions</h2>
+        <DataTable
+          checkAllRows={this.checkAllRows}
+          checkRow={this.toggleRowCheck}
+          headers={headers}
+          rows={schools}
+          showCheckboxes={true}
+          uncheckAllRows={this.uncheckAllRows}
+          uncheckRow={this.toggleRowCheck}
         />
-        <label>Created At:</label>
-        <span>{subscription.created_at}</span>
-        <h2>Payment Information</h2>
-        <label>Payment Method</label>
-        <ItemDropdown
-          callback={this.changePaymentMethod}
-          className="subscription-dropdown"
-          items={subscriptionPaymentOptions}
-          selectedItem={subscription.payment_method || 'N/A'}
-        />
-        <label>Purchase Amount (dollar value as integer -- no decimal or symbol)</label>
-        <input onChange={this.handlePaymentAmountChange} type="text" value={subscription.payment_amount / 100} />
-        <h2>Purchaser Information</h2>
-        {this.purchaserFromSchool()}
-        <label>Purchaser Email</label>
-        <p>If the purchaser is not in the school and you see a school dropdown, select &#39;None&#39; and put in the purchasers email.</p>
-        <input onChange={this.handlePurchaserEmailChange} type="text" value={subscription.purchaser_email} />
-        <br />
-        {this.changeToPurchaserInfo()}
-        {this.recurringIfCreditCard()}
-        <h2>Period {view === 'edit' ? <span className="warning text-red">-- These should not be edited without good reason!</span> : null}</h2>
+      </React.Fragment>
+    )
+  }
+
+  startDate = () => {
+    const { firstFocused, subscription } = this.state
+
+    return (
+      <React.Fragment>
         <label>Start Date</label>
         <p>
-          If this is a Teacher Subscription and no subscription already exists, the start date is set to today. If the subscription is being renewed, the start date is the day the old subscription ends.
+          If this is a Teacher Subscription and no subscription already exists,
+          the start date is set to today. If the subscription is being renewed,
+          the start date is the day the old subscription ends.
         </p>
+
         <SingleDatePicker
           date={subscription.start_date ? moment(subscription.start_date) : null}
           focused={firstFocused}
@@ -255,27 +330,121 @@ export default class EditOrCreateSubscription extends React.Component {
           onDateChange={this.handleStartDateChange}
           onFocusChange={() => this.setState({ firstFocused: !firstFocused })}
         />
-        <label htmlFor="">End Date</label>
-        <p>
-          If this a school's first paid subscription, the default end date is {promoExpiration}. This value just stated will update automatically depending on the time of year.
-        </p>
-        <SingleDatePicker
-          date={subscription.expiration ? moment(subscription.expiration) : null}
-          focused={secondFocused}
-          id="date-picker"
-          inputIconPosition="after"
-          navNext="›"
-          navPrev="‹"
-          numberOfMonths={1}
-          onDateChange={this.handleExpirationDateChange}
-          onFocusChange={() => this.setState({ secondFocused: !secondFocused })}
-        />
-        <div>
-          <button className="q-button cta-button bg-quillgreen text-white" onClick={submitAction} type="submit">
-            {view === 'new' ? 'New' : 'Update'} Subscription
-          </button>
-        </div>
+      </React.Fragment>
+
+    )
+  }
+
+  subscriptionInformation = () => {
+    const { subscription } = this.state
+
+    return (
+      <React.Fragment>
+        <h2>Subscription Information</h2>
+        {this.premiumStatus()}
+        {this.createdAt()}
+      </React.Fragment>
+    )
+  }
+
+  submitButton = () => {
+    const { subscriberType, view } = this.props
+    const submitAction = subscriberType === 'User' ? this.handleSubmit : this.submitConfirmation
+
+    return (
+      <div>
+        <button className="q-button cta-button bg-quillgreen text-white" onClick={submitAction} type="submit">
+          {view === 'new' ? 'New' : 'Update'} Subscription
+        </button>
       </div>
-    );
+    )
+  }
+
+  submitConfirmation = () => {
+    const { district, school } = this.props
+
+    let subscriberType = ''
+
+    if (district) {
+      subscriberType = 'District'
+    } else if (school) {
+      subscriberType = 'School'
+    }
+
+    if (confirm(`You know you are about to add/edit an entire ${subscriberType} subscription, right?`)) {
+      this.handleSubmit()
+    } else {
+      alert('submission canceled')
+    }
+  }
+
+  submitVars = () => {
+    const { subscription, schools } = this.state
+    const { view, subscriber, subscriberType } = this.props
+
+    const varsObj = {
+      data: {
+        subscription: subscription,
+        schools: schools,
+        authenticity_token: getAuthToken()
+      },
+      urlString: `${process.env.DEFAULT_URL}/cms/subscriptions`
+    }
+
+    if (view === 'edit') {
+      varsObj.httpVerb = 'put';
+      varsObj.urlString += `/${subscription.id}`;
+    } else if (view == 'new') {
+      varsObj.data.subscriber_id = subscriber.id
+      varsObj.data.subscriber_type = subscriberType
+      varsObj.httpVerb = 'post'
+    }
+    return varsObj
+  }
+
+  title = () => {
+    const { subscriber, view } = this.props
+
+    return <h1>{view === 'edit' ? 'Edit' : 'New'} Subscription: {_.get(subscriber, 'name')}</h1>
+  }
+
+  toggleRowCheck = (id) => {
+    const { schools } = this.state
+    const school = schools.find(school => school.id === id)
+    school.checked = !school.checked
+    this.setState({ schools })
+  }
+
+  checkAllRows = () => {
+    const { schools } = this.state
+    const newSchools = schools.map(school => {
+      school.checked = true
+      return school
+    })
+    this.setState({ schools: newSchools })
+  }
+
+  uncheckAllRows = () => {
+    const { schools } = this.state
+    const newSchools = schools.map(school => {
+      school.checked = false
+      return school
+    })
+    this.setState({ schools: newSchools })
+  }
+
+  render() {
+    return (
+      <div className="cms-subscription">
+        {this.title()}
+        {this.subscriptionInformation()}
+        {this.schoolSubscriptions()}
+        {this.paymentInformation()}
+        {this.purchaserInformation()}
+        {this.recurring()}
+        {this.period()}
+        {this.submitButton()}
+      </div>
+    )
   }
 }
