@@ -34,7 +34,7 @@ module Evidence
       options ||= {}
 
       super(options.reverse_merge(
-        only: [:id, :automl_model_id, :notes, :name, :labels, :state, :prompt_id, :created_at],
+        only: [:id, :automl_model_id, :notes, :name, :labels, :state, :prompt_id, :created_at, :updated_at],
         methods: [:older_models]
       ))
     end
