@@ -30,6 +30,8 @@ describe District, type: :model do
   it { should have_many(:district_admins) }
   it { should have_many(:admins).through(:district_admins) }
 
+  it_behaves_like 'a subscriber'
+
   context '#total_invoice' do
     let!(:district) { create(:district)}
     let!(:school) { create(:school)}
@@ -50,5 +52,4 @@ describe District, type: :model do
       expect(district.total_invoice).to eq((subscription.payment_amount + another_subscription.payment_amount) / 100)
     end
   end
-
 end
