@@ -54,13 +54,15 @@ require 'rails_helper'
 describe School, type: :model do
   it { should belong_to(:district) }
 
+  it_behaves_like 'a subscriber'
+
   let!(:bk_school) { create :school, name: "Brooklyn Charter School", zipcode: '11206'}
   let!(:queens_school) { create :school, name: "Queens Charter School", zipcode: '11385'}
   let!(:bk_teacher) { create(:teacher, school: bk_school) }
   let!(:bk_teacher_colleague) { create(:teacher, school: bk_school) }
   let!(:queens_teacher) { create(:teacher, school: queens_school) }
 
-  describe('#subscription') do
+  describe '#subscription' do
     let!(:subscription) { create(:subscription, expiration: Date.tomorrow) }
     let!(:school_subscription) {create(:school_subscription, school: bk_school, subscription: subscription)}
 
@@ -80,7 +82,7 @@ describe School, type: :model do
     end
   end
 
-  describe('#present_and_future_subscriptions') do
+  describe '#present_and_future_subscriptions' do
     let!(:subscription) { create(:subscription, expiration: Date.tomorrow) }
     let!(:next_subscription) { create(:subscription, expiration: Date.tomorrow + 1.year, start_date: Date.tomorrow) }
     let!(:school_subscription) {create(:school_subscription, school: bk_school, subscription: subscription)}
@@ -102,56 +104,53 @@ describe School, type: :model do
     end
   end
 
-
   describe 'validations' do
-    before do
-      @school = School.new
-    end
+    let(:school) { School.new }
 
     it 'lower grade is within bounds' do
-      @school.lower_grade = 5
+      school.lower_grade = 5
 
-      expect(@school.valid?).to eq(true)
+      expect(school).to be_valid
     end
 
     it 'lower grade is out of bounds' do
-      @school.lower_grade = -1
+      school.lower_grade = -1
 
-      expect(@school.valid?).to eq(false)
-      expect(@school.errors[:lower_grade]).to eq(['must be between 0 and 12'])
+      expect(school).not_to be_valid
+      expect(school.errors[:lower_grade]).to eq(['must be between 0 and 12'])
     end
 
     it 'upper grade is within bounds' do
-      @school.upper_grade = 8
+      school.upper_grade = 8
 
-      expect(@school.valid?).to eq(true)
+      expect(school).to be_valid
     end
 
     it 'upper grade is out of bounds' do
-      @school.upper_grade = 14
+      school.upper_grade = 14
 
-      expect(@school.valid?).to eq(false)
-      expect(@school.errors[:upper_grade]).to eq(['must be between 0 and 12'])
+      expect(school).not_to be_valid
+      expect(school.errors[:upper_grade]).to eq(['must be between 0 and 12'])
     end
 
     it 'lower grade is below upper grade' do
-      @school.lower_grade = 2
-      @school.upper_grade = 8
+      school.lower_grade = 2
+      school.upper_grade = 8
 
-      expect(@school.valid?).to eq(true)
+      expect(school).to be_valid
     end
 
     it 'lower grade is above upper grade' do
-      @school.lower_grade = 6
-      @school.upper_grade = 3
+      school.lower_grade = 6
+      school.upper_grade = 3
 
-      expect(@school.valid?).to eq(false)
-      expect(@school.errors[:lower_grade]).to eq(['must be less than or equal to upper grade'])
+      expect(school).not_to be_valid
+      expect(school.errors[:lower_grade]).to eq(['must be less than or equal to upper grade'])
     end
 
   end
 
-  describe('school_year_start method') do
+  describe 'school_year_start method' do
     it 'fetches 07-01 of this year if the date is after 07-01' do
       time = Date.parse('2020-07-01')
       expect(School.school_year_start(time)).to eq(time.beginning_of_day)
@@ -163,33 +162,48 @@ describe School, type: :model do
     end
   end
 
-  describe('district admin behavior') do
+  describe 'district admin behavior' do
     let(:school) { create(:school) }
     let(:district) { create(:district) }
+    let(:another_district) { create(:district) }
     let(:admin) { create(:user)}
 
-    it 'creates new school admin record if a school is attached to a new district' do
-      create(:district_admin, user: admin, district: district)
-      expect(SchoolsAdmins.find_by(school: school, user: admin)).not_to be
-      school.update(district_id: district.id)
-      expect(SchoolsAdmins.find_by(school: school, user: admin)).to be
+    before { create(:district_admin, user: admin, district: district) }
+
+    describe '#attach_new_district_school_admins' do
+      it { expect { school }.to not_change(SchoolsAdmins, :count) }
+
+      it 'creates school admin if a school is attached to a new district' do
+        expect(SchoolsAdmins.find_by(school: school, user: admin)).not_to be
+        school.update(district: district)
+        expect(SchoolsAdmins.find_by(school: school, user: admin)).to be
+      end
+
+      context 'district admin is already admin for school' do
+        before { create(:schools_admins, user: admin, school: school) }
+
+        it {  expect { school.update(district: district) }.to not_change(SchoolsAdmins, :count) }
+      end
     end
 
-    it 'does not create new school admin record if a school is attached to a new district and the district admin is already admin for the school' do
-      create(:schools_admins, user: admin, school: school)
-      expect(SchoolsAdmins.where(school: school, user: admin).count).to eq 1
+    describe '#detach_old_district_school_admins' do
+      context 'school has no district' do
+        before { school }
 
-      school.update(district_id: district.id)
-      expect(SchoolsAdmins.where(school: school, user: admin).count).to eq 1
-    end
+        it { expect { school.update(district: nil) }.to not_change(SchoolsAdmins, :count) }
+      end
 
-    it 'destroys school admin record if a school is detached from a district' do
-      create(:district_admin, user: admin, district: district)
-      school.update(district_id: district.id)
-      expect(SchoolsAdmins.find_by(school: school, user: admin)).to be
+      context 'school is attached to district' do
+        before { school.update(district: district) }
 
-      school.update(district_id: nil)
-      expect(SchoolsAdmins.find_by(school: school, user: admin)).not_to be
+        it { expect { school.update(district: nil) }.to change(SchoolsAdmins, :count).from(1).to(0) }
+      end
+
+      context 'school is attached to another district' do
+        before { school.update(district: another_district) }
+
+        it { expect { school.update(district: nil) }.to not_change(SchoolsAdmins, :count) }
+      end
     end
   end
 end
