@@ -29,7 +29,6 @@
 #  index_subscriptions_on_stripe_invoice_id  (stripe_invoice_id) UNIQUE
 #
 require 'rails_helper'
-require 'ostruct'
 
 describe Subscription, type: :model do
   context 'validations' do
@@ -61,7 +60,7 @@ describe Subscription, type: :model do
 
   describe 'the subscription type consts' do
     it "the official school and offical teacher types contain the same values as all offical types" do
-      types_by_role = Subscription::OFFICIAL_SCHOOL_TYPES.dup.concat(Subscription::OFFICIAL_TEACHER_TYPES)
+      types_by_role = Subscription::OFFICIAL_DISTRICT_TYPES + Subscription::OFFICIAL_SCHOOL_TYPES + Subscription::OFFICIAL_TEACHER_TYPES
       expect(types_by_role.uniq).to match_array(Subscription::ALL_OFFICIAL_TYPES)
     end
   end
@@ -127,94 +126,6 @@ describe Subscription, type: :model do
     end
   end
 
-  describe ".give_school_premium_if_charge_succeeds" do
-    let!(:school) { create(:school) }
-    let!(:user) { create(:user) }
-    let!(:subscription) { build(:subscription, expiration: Date.new(2018,4,6), purchaser: user, account_type: 'School Paid') }
-    # let!(:school_sub) {build(:school_subscription, school: school, subscription: subscription)}
-
-    before do
-      allow_any_instance_of(Subscription).to receive(:charge_user_for_school_premium).and_return({status: 'succeeded'})
-    end
-
-    it "calls #Subscription.new_school_premium_sub" do
-      expect(Subscription).to receive(:new_school_premium_sub).with(school, user).and_return(subscription)
-      Subscription.give_school_premium_if_charge_succeeds(school, user)
-    end
-
-    it "calls #Subscription.save_if_charge_succeeds" do
-      expect_any_instance_of(Subscription).to receive(:save_if_charge_succeeds)
-      Subscription.give_school_premium_if_charge_succeeds(school, user)
-    end
-
-    it "creates a new subscription when the charge succeeds" do
-      old_sub_count = Subscription.count
-      Subscription.give_school_premium_if_charge_succeeds(school, user)
-      expect(Subscription.count - old_sub_count).to eq(1)
-    end
-
-    it "creates a new user-subscription join when the charge succeeds" do
-      # school.users.push(user)
-      old_user_sub_count = UserSubscription.count
-      subby = Subscription.give_school_premium_if_charge_succeeds(school, user)
-      expect(SchoolSubscription.count - old_user_sub_count).to eq(1)
-    end
-
-    it "creates a new subscription with the correct payment amount" do
-      new_sub = Subscription.give_school_premium_if_charge_succeeds(school, user)
-      expect(new_sub.payment_amount).to eq(Subscription::SCHOOL_FIRST_PURCHASE_PRICE)
-    end
-
-    it "creates a new subscription with the correct payment method" do
-      new_sub = Subscription.give_school_premium_if_charge_succeeds(school, user)
-      expect(new_sub.payment_method).to eq('Credit Card')
-    end
-
-    it "creates a new subscription with the correct contact" do
-      new_sub = Subscription.give_school_premium_if_charge_succeeds(school, user)
-      expect(new_sub.purchaser).to eq(user)
-    end
-
-    context 'when the charge does not suceed' do
-      before do
-        allow_any_instance_of(Subscription).to receive(:charge_user_for_school_premium).and_return({status: 'failed'})
-      end
-
-      it "does not create a new subscription" do
-        old_sub_count = Subscription.count
-        Subscription.give_school_premium_if_charge_succeeds(school, user)
-        expect(Subscription.count - old_sub_count).to eq(0)
-      end
-
-      it "does not create a new school-subscription join" do
-        old_school_sub_count = UserSubscription.count
-        Subscription.give_school_premium_if_charge_succeeds(school, user)
-        expect(SchoolSubscription.count - old_school_sub_count).to eq(0)
-      end
-
-    end
-  end
-
-  describe ".school_or_user_has_ever_paid?" do
-    let!(:subscription) { create(:subscription) }
-    let!(:user) { create(:user) }
-    let!(:user_subscription) { create(:user_subscription, subscription: subscription, user: user) }
-
-    it "responds with true if school or user has ever had anything in the OFFICIAL_PAID_TYPES_LIST" do
-      Subscription::OFFICIAL_PAID_TYPES.each do |type|
-        subscription.update(account_type: type)
-        expect(Subscription.school_or_user_has_ever_paid?(user)).to be
-      end
-    end
-
-    it "responds with false if school or user has only had things in the OFFICIAL_FREE_TYPES_LIST" do
-      Subscription::OFFICIAL_FREE_TYPES.each do |type|
-        subscription.update(account_type: type)
-        expect(Subscription.school_or_user_has_ever_paid?(user)).not_to be
-      end
-    end
-  end
-
   describe ".redemption_start_date" do
     let!(:school) { create(:school) }
     let!(:subscription) { create(:subscription, expiration: Date.tomorrow) }
@@ -253,13 +164,15 @@ describe Subscription, type: :model do
     end
   end
 
-  describe 'create_with_user_join' do
+  describe 'create_and_attach_subscriber' do
     let!(:user) { create(:user) }
-    let(:old_sub) { Subscription.create_with_user_join(user.id, expiration: Date.yesterday, account_type: 'Teacher Paid') }
+    let(:old_sub) do
+      Subscription.create_and_attach_subscriber({ expiration: Date.yesterday, account_type: 'Teacher Paid' }, user)
+    end
 
     it 'creates a subscription based off of the passed attributes' do
       attributes = { expiration: Date.yesterday, account_type: 'Teacher Paid' }
-      new_sub = Subscription.create_with_user_join(user.id, attributes)
+      new_sub = Subscription.create_and_attach_subscriber(attributes, user)
       expect(new_sub.account_type).to eq('Teacher Paid')
       expect(new_sub.expiration).to eq(Date.yesterday)
     end
@@ -268,7 +181,7 @@ describe Subscription, type: :model do
 
       it 'adds at least a year (or more, depending on promotions) to other accounts' do
         attributes = { account_type: 'Teacher Paid' }
-        new_sub = Subscription.create_with_user_join(user.id, attributes)
+        new_sub = Subscription.create_and_attach_subscriber(attributes, user)
         expect(new_sub.expiration).to be >= 365.days.from_now.to_date
       end
     end
@@ -277,46 +190,44 @@ describe Subscription, type: :model do
       let!(:user_with_existing_subscription) { create(:user) }
 
       let!(:existing_subscription ) do
-        Subscription.create_with_user_join(
-          user_with_existing_subscription.id,
-          expiration: 365.days.from_now.to_date,
-          account_type: 'Teacher Paid'
-       )
+        Subscription.create_and_attach_subscriber(
+          { expiration: 365.days.from_now.to_date, account_type: 'Teacher Paid' },
+          user_with_existing_subscription
+        )
       end
 
       let!(:user_with_no_existing_subscription) { create(:user) }
       let!(:user_with_expired_existing_subscription) { create(:user) }
 
       let!(:existing_expired_subscription ) do
-        Subscription.create_with_user_join(
-          user_with_expired_existing_subscription.id,
-          expiration: 1.day.ago.to_date,
-          account_type: 'Teacher Paid'
+        Subscription.create_and_attach_subscriber(
+          { expiration: 1.day.ago.to_date, account_type: 'Teacher Paid' },
+          user_with_expired_existing_subscription
         )
       end
 
       it 'creates a trial subscription with an expiration in 30 days if there is no existing subscription' do
         attributes = { account_type: 'Teacher Trial' }
-        new_sub = Subscription.create_with_user_join(user_with_no_existing_subscription.id, attributes)
+        new_sub = Subscription.create_and_attach_subscriber(attributes, user_with_no_existing_subscription)
         expect(new_sub.expiration).to eq 30.days.from_now.to_date
       end
 
       it 'creates a trial subscription with an expiration in 30 days if there is an expired existing subscription' do
         attributes = { account_type: 'Teacher Trial' }
-        new_sub = Subscription.create_with_user_join(user_with_expired_existing_subscription.id, attributes)
+        new_sub = Subscription.create_and_attach_subscriber(attributes, user_with_expired_existing_subscription)
         expect(new_sub.expiration).to eq 30.days.from_now.to_date
       end
 
       it 'create a trial subscription with an expiration 31 days after the existing subscription expiration' do
         attributes = { account_type: 'Teacher Trial' }
-        new_sub = Subscription.create_with_user_join(user_with_existing_subscription.id, attributes)
+        new_sub = Subscription.create_and_attach_subscriber(attributes, user_with_existing_subscription)
         expect(new_sub.expiration).to eq(existing_subscription.expiration + 31)
       end
     end
 
     it 'makes a matching UserSubscription join' do
       attributes = { expiration: Date.yesterday, account_type: 'Teacher Paid' }
-      new_sub = Subscription.create_with_user_join(user.id, attributes)
+      new_sub = Subscription.create_and_attach_subscriber(attributes, user)
       join = new_sub.user_subscriptions.first
       expect([join.user_id, join.subscription_id]).to eq([user.id, new_sub.id])
     end
@@ -324,7 +235,7 @@ describe Subscription, type: :model do
     context 'when the subscription already exists' do
       it 'updates a UserSubscription based off of the passed attributes' do
         attributes = { expiration: Date.tomorrow }
-        Subscription.create_with_user_join(user.id, attributes)
+        Subscription.create_and_attach_subscriber(attributes, user)
         expect(user.reload.subscription.expiration).to eq(Date.tomorrow)
       end
     end
@@ -411,22 +322,6 @@ describe Subscription, type: :model do
           expect(Stripe::Subscription).to receive(:create)
           subject
         end
-      end
-    end
-
-    describe '#renewal_price' do
-      let!(:school) { create(:school) }
-      let!(:school_subscription) { create(:school_subscription, subscription: subscription, school: school) }
-
-      subject { subscription.renewal_price }
-
-      it "returns the school renewal price if any schools are associated with the subscription" do
-        expect(subject).to eq(Subscription::SCHOOL_RENEWAL_PRICE)
-      end
-
-      it "returns the teacher renewal price if no schools are associated with the subscription" do
-        school_subscription.destroy
-        expect(subject).to eq(Subscription::TEACHER_PRICE)
       end
     end
 
